Update all import paths to include file extensions
I'm not crazy about this change, as I liked it when it inferred the extension automatically.

Also, it feels weird to provide a .js file extension when I'm writing TypeScript files.

However, this seems to be necessary for pure-ESModule packages, and after installing `markdown-table` I need to make these changes.

[1]:https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c#faq
[2]:https://github.com/microsoft/TypeScript/issues/42813
[3]:https://www.typescriptlang.org/docs/handbook/module-resolution.html#classic
[4]:https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules