Exclude empty co-authors in commits
Previously, we used the `*` regular expression token, which matched
any characters following "Co-authors: " zero to unlimited times.

This would allow commits that appended an empty "Co-authors: " string
with no co-authors defined to show up in the pairing data, which would
result in an empty row and column in the generated matrix.

This changes it to use the `+` token, which matches one to unlimited times.
So, this will only include commits if at least one character is included
following "Co-authors: ".